Add python hello world
In this commit, the README has been changed to suite more for the public and not just for me playing around.

Also a python directory/script has been added, to mimic more a real project. It's still an Hello world application

Interesting discovery is that when I change the branch in the desktop app and click of fetsh origin. The new "Desktop-branch" get's loaded in to GetHub.
However it doesn't load the changes that have not been commited. So it just looked like another master branch ( since that is copied) on the GitHub.
While here at the desktop app I have a nice overview on what has been changed.