Add C++ Hello world
After commiting the Python add. It was commited localy on the computer. Since it hasn't been updated on the site -thus the commit was just here on the computer and not an the server.

For this reason i'm making a seccond commit (just and hello world c++ file). This is to test what would happen if I make two commits, localy and then upload it to the server